Add custom xml attributes to ibb data element
It might not be the best practice to add custom attribute
to the existing elements. It is suggested to pass in a custom
element with required attributes and data, but unlike 'message
stanza with 'iq' that does not seem to work.

Because of that passing in two new attributes 'nn-comment'
(useful to add comments to image) and 'nn-message-group'
(useful to group images when multiple files shared). It is not
the stanza element(iq), but inner data element which is passed
on with this new attributes.

See #17.
See #18.